fix: Makefile working (make lib and use this lib)/ add: struct (prepare for generic use in future)/ fix: all include with new makefile juste need #include "my.hpp" and not presise include/

+++ level.cpp
@@ -5,7 +5,7 @@
 ** main
 */
 
-#include "include/my.hpp"
+#include "my.hpp"
 
 int max_length(std::vector<std::string> map)
 {
@@ -13,7 +13,7 @@ int max_length(std::vector<std::string> map)
     int testing = 0;
 
     for (size_t i = 0; i < map.size(); i++) {
-        for (int j = 0; j < map[i].size(); j++)
+        for (size_t j = 0; j < map[i].size(); j++)
             testing++;
         if (testing > max)
             max = testing;
